+++ packages/pi-goal/src/index.ts
@@ -9,5 +9,6 @@ export * from "./continuation.js";
 export * from "./tools.js";
 export * from "./ui.js";
 export * from "./rendering.js";
+export * from "./metadata.js";
 export * from "./install-telemetry.js";
 export * from "./provider-limits.js";